+++ hc/accounts/tests/test_profile.py
@@ -83,6 +83,10 @@ class ProfileTestCase(BaseTestCase):
         r = self.client.get("/accounts/profile/")
         self.assertContains(r, "Alices Key")
 
+        # It should show a warning about Alices Key being the only second factor
+        s = """The key "Alices Key" is currently your only second factor."""
+        self.assertContains(r, s)
+
     def test_it_handles_unusable_password(self):
         self.alice.set_unusable_password()
         self.alice.save()
@@ -106,3 +110,38 @@ class ProfileTestCase(BaseTestCase):
         self.assertContains(r, "Enabled")
         self.assertContains(r, "configured on Jan 1, 2020")
         self.assertNotContains(r, "Set Up Authenticator App")
+
+        # It should show a warning about TOTP being the only second factor
+        s = "The Authenticator app is currently your only second factor."
+        self.assertContains(r, s)
+        self.assertContains(r, "or register a Security Key to be used")
+
+    def test_it_shows_no_warning_if_multiple_keys_are_registered(self):
+        Credential.objects.create(user=self.alice, name="Alices Key")
+        Credential.objects.create(user=self.alice, name="Alices Other Key")
+
+        self.client.login(username="alice@example.org", password="password")
+        r = self.client.get("/accounts/profile/")
+
+        self.assertNotContains(r, "is currently your only second factor.")
+
+    def test_it_shows_no_warning_if_key_and_totp_is_registered(self):
+        Credential.objects.create(user=self.alice, name="Alices Key")
+        self.profile.totp = "0" * 32
+        self.profile.totp_created = "2020-01-01T00:00:00+00:00"
+        self.profile.save()
+
+        self.client.login(username="alice@example.org", password="password")
+        r = self.client.get("/accounts/profile/")
+
+        self.assertNotContains(r, "is currently your only second factor.")
+
+    @override_settings(RP_ID=None)
+    def test_it_does_not_mention_security_key_if_rp_id_is_not_set(self):
+        self.profile.totp = "0" * 32
+        self.profile.totp_created = "2020-01-01T00:00:00+00:00"
+        self.profile.save()
+
+        self.client.login(username="alice@example.org", password="password")
+        r = self.client.get("/accounts/profile/")
+        self.assertNotContains(r, "or register a Security Key to be used")